Address comments for tests
