+++ src/client/Microsoft.Identity.Client/AppConfig/ManagedIdentityApplicationBuilder.cs
@@ -45,7 +45,7 @@ namespace Microsoft.Identity.Client
         {
             ApplicationBase.GuardMobileFrameworks();
 
-            var config = new ApplicationConfiguration(isConfidentialClient: true);
+            var config = new ApplicationConfiguration(isConfidentialClient: false);
             var builder = new ManagedIdentityApplicationBuilder(config).WithOptions(options);
 
             if (!string.IsNullOrWhiteSpace(options.UserAssignedClientId))
@@ -95,7 +95,7 @@ namespace Microsoft.Identity.Client
                 throw new ArgumentNullException(nameof(userAssignedId));
             }
 
-            var config = new ApplicationConfiguration(isConfidentialClient: true);
+            var config = new ApplicationConfiguration(isConfidentialClient: false);
             return new ManagedIdentityApplicationBuilder(config)
                 .WithUserAssignedManagedIdentity(userAssignedId)
                 .WithCacheSynchronization(false);
